Began the typescript conversion process
