+++ package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "start-session-service": "nodemon ./services/session-service/main.js",
     "start-items-service": "nodemon ./services/items-service/main.js",
-    "start-user-service": "nodemon ./services/user-service/main.js"
+    "start-user-service": "nodemon ./services/user-service/main.ts"
   },
   "dependencies": {
     "@grpc/proto-loader": "0.5.3",
@@ -14,18 +14,24 @@
     "compression": "1.7.4",
     "dotenv": "8.2.0",
     "express": "4.17.1",
-    "express-validator": "6.3.1",
-    "google-protobuf": "3.11.2",
+    "express-validator": "6.4.0",
+    "google-protobuf": "3.11.3",
     "grpc": "1.24.2",
     "http-status-codes": "1.4.0",
     "jsonwebtoken": "8.5.1",
     "lodash": "4.17.15",
     "morgan": "1.9.1",
-    "pg": "7.17.1",
+    "pg": "7.18.1",
     "redis": "2.8.0",
+    "ts-custom-error": "3.1.1",
+    "ts-node": "^8.6.2",
+    "typescript": "3.7.5",
     "winston": "3.2.1"
   },
   "devDependencies": {
+    "@types/express": "4.17.2",
+    "@types/morgan": "1.7.37",
+    "@types/winston": "2.4.4",
     "nodemon": "2.0.2"
   },
   "license": "ISC"
